third_party/blink/web_tests/external/wpt/webvtt/api/VTTCue/getCueAsHTML-expected.txt

This is a testharness.js-based test.
[FAIL] VTTCue.getCueAsHTML(), x\\0
  assert_equals: data expected "x\\0" but got "x"
Harness: the test ran to completion.
